server: rename modules to kebab-case
