+++ src/server/event-stream-handler.js
